+++ Code/glimac/include/project/StaticObject.hpp
@@ -6,7 +6,7 @@ class StaticObject : public Object {
 public:
 
     // constructor 
-    StaticObject(char, int, int, float, float, Orientation);
+    StaticObject(char, float, float, float, float, Orientation);
     StaticObject();
 
     void display() override;
